+++ node/ts/Errors.ts
@@ -5,6 +5,11 @@
 
 import { ProtocolAddress, ServiceId } from './Address.js';
 import * as Native from './Native.js';
+import { newNativeHandle } from './internal.js';
+import {
+  convertNativeRegistrationSessionState,
+  type RegistrationSessionState,
+} from './net/RegistrationSession.js';
 
 export enum ErrorCode {
   Generic,
@@ -132,6 +137,7 @@ export class LibSignalErrorBase extends Error {
   public readonly code: ErrorCode;
   public readonly operation: string;
   readonly _addr?: string | Native.ProtocolAddress;
+  readonly _sessionState?: Native.RegistrationSession;
 
   constructor(
     message: string,
@@ -171,6 +177,15 @@ export class LibSignalErrorBase extends Error {
       default:
         throw new TypeError(`cannot get address from this error (${this})`);
     }
+  }
+
+  public get sessionState(): RegistrationSessionState | undefined {
+    if (this._sessionState === undefined) {
+      return undefined;
+    }
+    return convertNativeRegistrationSessionState(
+      newNativeHandle(this._sessionState)
+    );
   }
 
   public toString(): string {
@@ -427,10 +442,12 @@ export type RegistrationRequestRejectedError = LibSignalErrorCommon & {
 export type RegistrationSessionNotReadyForVerificationError =
   LibSignalErrorCommon & {
     code: ErrorCode.RegistrationSessionNotReadyForVerification;
+    readonly sessionState?: RegistrationSessionState;
   };
 
 export type RegistrationVerificationSendFailedError = LibSignalErrorCommon & {
   code: ErrorCode.RegistrationVerificationSendFailed;
+  readonly sessionState?: RegistrationSessionState;
 };
 
 export type RegistrationVerificationCodeNotDeliverableError =